Fix Fur-76 Advanced Search - Date Picker with a wrapped-up DatePicker component to reduce the code duplicate.
