+++ src/utils/allowedOrigins.js
@@ -1,6 +1,6 @@
 const allowedOrigins = [
   // 'https://www.yoursite.com',
-  // 'http://localhost:3000' // default React dev server port
+  'http://localhost:3000', // default React dev server port
   'http://localhost:3001',
   'https://furfinder.onrender.com'
 ];
